Break key Components out into separeate files, separating not only implemenation code but also CSS and resources.
