+++ src/spas/home/components/Score.font.js
@@ -1,8 +1,8 @@
-const BUNDLE_HELPER = require('../../utils')
+const BUNDLE_HELPER = require('../../../utils')
 
 module.exports = {
     files: [
-      './icons/beer.svg'
+      '../icons/beer.svg'
     ],
     fileName: BUNDLE_HELPER.path(__dirname) + '/fonts.[ext]'
   }
