+++ src/spas/home/components/Target.css
@@ -1,0 +1,8 @@
+.target
+{
+	fill: pink;
+	width: 100px;
+	height: 100px;
+	position: absolute;
+	animation:turn 3s linear infinite normal;
+}
